Deps: Move out to separate TU
