+++ src/classes/resources.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <util.h>
+#include <wren.h>
 
 // foreign method binding
